feat: add warnings chip
