+++ app/components/WarningsButton.vue
@@ -1,5 +1,20 @@
 <template>
+    <UChip
+        v-if="warningsCount > 0"
+        inset
+        color="yellow"
+        position="bottom-right"
+    >
+        <UButton
+            :icon="'i-heroicons-exclamation-triangle-20-solid'"
+            color="gray"
+            variant="ghost"
+            aria-label="Warnings"
+            @click="openSlideover"
+        />
+    </UChip>
     <UButton
+        v-else
         :icon="'i-heroicons-exclamation-triangle-20-solid'"
         color="gray"
         variant="ghost"
@@ -12,6 +27,8 @@
 import WarningsSlideover from "./WarningsSlideover.vue";
 
 const slideover = useSlideover();
+const { warningsCount } = useWarnings();
+
 function openSlideover() {
     slideover.open(WarningsSlideover);
 }
